BUG FIX - Uncaught exception when accessing admin with none existent user.
Summary:

When attempting to login to the Magento admin. If the user does not exist the resource method loadByUsername returns false. This result is passed to setData that will then throw an exception.

Fix:

Added a test to confirm $user is returned before calling setData. This allows the execution to correctly fail through to the standard exception handler and display an appropriately secure exception message.