Adding dispatch of an event before and after applyAllUpdates().
This allows for the ability to put in a mechanism to prevent errors
being displayed to users if more than one make a request in the same
span of time that the install/upgrade scripts are running.  

This mechanism could be: waiting, sending to an seperate page, etc.  

Now, we would not have to turn on maintainence mode...potentially.