Add events to transactional email sending
The purpose of the events is to enable modules to
- add attachments
- change email templates
- modify template variables
- add template variables
- modify email content

Also, the patch makes the mailer class factory method configurable.
The purpose is to enable modules to use a diferent mailer without
reverting to a rewrite of Mage_Core_Model_Email_Template.

Additionaly the undeclared dependency on Mage_Newsletter was
removed from Mage_Core_Model_Email_Template, and an event observer
added to the newsletter module taking care the business logic
is not changed.