Update system.xsd
Added in A-Z in to the typeModel because if you don't this doesn't allow CaMeL case in the company name when you're defining you're own source models which are not present in Magento eg payment actions for Payment Gateways. eg my Module company directory started JuicyMedia but the regex was preventing that.