This fix is for Stores with large amounts of Products and Configurable Attributes. _loadSkuSuperAttributeValues() is given parameters and a foreach loop creates the productIds needed for the current bunch.
The conversion From Sku to Id is duplicated later in save(), but I felt it is best to do the conversion here then let the DB do the work.

This fix is based off of a module rewrite for EE 1.12 and has not been tested on Mage2. It should work though, That Module successfully imported 700,000 Simple Products and many Configurable Products. Most of the Configurable Products having 3 super_attributes.

Signed-off-by: Steven Hoffman <s.hoffman@harapartners.com>