Add Failured-Allowed PHP 5.6 testing to Travis
While PHP 5.6 is not officially supported by Magento proper, allowing
Magento 2 to be installed and tested under that environment will allow
tests to be constructed so that features or functionality broken in these
versions may be understood and potentially fixed.  If the tests are
failing for these environments it also warns developers of the
reprecussions of using Magento in them.