Make Browser configurable with command line option
Move the check if the browser binary exists into the block with the default
discovery. This makes it possible to configure the Browser in the jsTestDriver.php
file with command line arguments.
For example:

'Browser' => '/Applications/Firefox.app/Contents/MacOS/firefox -P "selenium"',

Without the change the check would fail, since the value is not an existing
file name, but rather a file name including command line options.