Refactor run_js_tests.php
All global functions in the CLI script have been moved into a new class JsTestDriver.php
located within the dev/tests/js/framework directory.
The script itself is now just an entry point to call functionality within the new class.

Purpose of this refactoring is to

* Make the code more reusable and maintainable
* Make the script work on OS X (without killing PHPStorm)
* Provide an way to rebuild the jsTestDriver.conf without running the tests

The public interface of the JsTestDriver class is:

* JsTestRunner::fromConfigFile() - static factory method
* runTests() - functionality formerly within the run_js_tests.php script
* buildJsTestRunnerConfig() - rebuild the jsTestDriver.conf file

The functionality of the script has been decomposed into smaller functions.

A new entry point script has been added, build_js_test_conf.php, to just
rebuild the test configuration, without executing JsTestDriver.

This should be considered not the final solution, but rather a step on the way
to a flexible and cross-plattform way to run the js tests.