Replace UTF8 'en dash' with minus in error msg
The character in the error message was the UTF8 character – U+2013 (&\#8211)
also called "en dash", which is used e.g. to indicate a range of values.

For command line options the character - U+002D	(&\#45) is used to specify
command options. This change should allow users who see the error message
to copy the command out of the browser and paste it into the shell.