Apply pattern matching datasource config files
The Magento\Ui\DataProvider\Config\FileResolver class currently
uses the hardcoded pattern '/*/*/etc/data_source/*' to search
for data source configuration files.
The method argument $filename, which contains the file pattern that should
be matched, is not used.
Due to this temorarily excludng files by renaming them
with a different file type ending is not possible.
For example

    etc/data_source/customer.xml -> etc/data_source/customer.xml.off

This patch applies the specified file name pattern to the pattern.