MAGETWO-35060: Copy and Paste detector is run against test files that are blacklisted
The blacklist file has paths that are relative to the white list and to BP,
as a result only half the blacklist can ever be honored.  Previous commits
fixed the testCopyPaste test to use the whitelist like all other tests
in LiveCodeTest class.

Since this change has caused some confusion and since there is already
evidence of confusion on how to use the blacklist for copy paste detection,
this commit will revert functionality back to the old behavior of using BP.

In the future the whitelist should be used consistently for all tests in the
LiveCodeTest class and the blacklist should be cleaned up.