Rename \Magento\Framework\Api\SearchCriteriaBuilder::addFilter() to addFilters()
Purpose
==

Make method name more descriptive and easier to understand.

Background
==

Having a method called `addFilter()` which take an array of `Filter` instances caused
many doubletakes and facepalms during the m2 dev training.
It seems as if the method name originated from the old collection `addFieldToFilter`
from collections in Magento 1, where the condition type was set as a key in the passed
argument array (e.g. `addFieldToFilter('name', ['eq' => $filter])`).
In fact, some integration tests even set such expectations on the searchCriteriaBuilder,
for example *app/code/Magento/Sales/Test/Unit/Model/Service/CreditmemoServiceTest.php*.

However, this must be a bug because the array keys in the argument array are ignored,
and the code only happened to work because the repositories in question default to
using the 'eq' operator when applying the filter groups to the collection if no condition
type is set on the filter instance.

Reasons for this PR
==

Make the method name more intuitive.
Naming definitely could still be improved, but at least it no longer is that misleading.
Maybe using `addFilterGroupWithFilters()` would be even more descriptive, but I chose the
smaller change, since it probably is "good enough".

Remove wrong examples of method usage from tests, so they don't serve as wrong
documentation.