Change misleading class method name
In Magento\Framework\Filesystem\DirectoryList#L130 the
private method has a name that doesn't reflect what the
method actually does. It does not filter paths, but it
perform a conversion (as the doc block suggests as well).

This commit changes the method name from 'filterPath' to
'convertPathSlashes' to improve code clarity and understanding.