string false does not evaluate to false
The false value used here https://github.com/magento/magento2/blob/develop/app/code/Magento/Sales/etc/pdf.xml#L33 gets represented as a string which leads to all totals being shown in the pdfs even if the amount is 0 and it is configured to not show. Checking for the string true fixes it.