Improve \Magento\Framework\Api\SortOrder
Purpose
==

Make the specification of sort directions on sort criteria easier for developers.

Background
==

This PR addresses several issues:

**PHPDoc type hint mismatch**
All parameter type hints expecting a sort direction specify `@param string $direction`,
but the constants `SearchCriteriaInterface::SORT_ASC` and `SearchCriteriaInterface::SORT_DESC`
are integers.
As a developer, when I see a method `setDirection()` that takes a string, I expect the
argument to be either `ASC` or `DESC`. Neither work, but there is no error helpful error
message.

**Unintuitive placement of constants**
Also the location of the constants is not intuitive. When working with the code in question,
a developer is focused on the classes SortOrder and SortOrderBuilder. So the expectation
would be to find the class constants on either of those classes.

**Missing validation**
The sort direction can be specified as a scalar argument, but there is no validation.
This PR adds validation and throws an explanatory exception on validation failure.

Reasons for this PR
==

The motivation is to make it easier and more intuitive for developers to specify a sort order.