increase length of sales_order_payment:last_trans_id from 32 to 100
txn_id field in sales_payment_transaction table is length 100, since last_trans_id in the sales_order_payment table links directly to that field both should be of the same type and length. Personally I found out as the transaction IDs of the payment gateway I'm trying to integrate is more than 32 characters long.

I'm not sure if this needs to be added on an db update script at this point.