Add improvements for varnish vcl files
Hi,

I have created this PR to improve the varnish configuration a little bit and fix some issues.
I don't have varnish 3 at the moment, so I did not add the same configuration to the Varnish 3
files, because I can't test it.

This PR is for https://github.com/magento/magento2/issues/1575