Clean up class and interface dependencies on `ComponentRegistrar`
Summary:
* Move the `ComponentRegistrar` type constants to `ComponentRegistrarInterface`
* Have clients of `ComponentRegistrar` depend on the interface only where possible
* This PR introduces **no breaking changes**, it contains just some cleanup refactoring.

Since the methods declared in the interface require the type code as part of the argument list, it makes more
sense in my view to have the type code constants be part of the interface declaration,
rather then the concrete implementation.
Because PHP allows using interface constants on classes implementing that interface this is not a breaking change.
For example, any reference to `ComponentRegistrar::MODULE` will continue to resolve cleanly, even though now
the constant is declared as `ComponentRegistrarInterface::MODULE`.

With this change in place, many files dependent on both `\Magento\Framework\Component\ComponentRegistrar`
and `\Magento\Framework\Component\ComponentRegistrarInterface` can now only depend on the interface.

Because of this the second part of the PR is simply cleaning up the dependencies where possible.
Any class instantiating the `ComponentRegistrar` directly remains unchanged, but any class requiring
it as a dependency can now only depend on `ComponentRegistrarInterface`.