Make email subject translatable
When the email templates are being processed the template subject is not
translatable because the translation happens within the filters via
'trans'. We must add the translated subject before the processor since
there the variables used in the subject will be replaced and thus no
longer be generic for translation.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>