FQCN do not start with a backslash
there must have been a misconception by the developer who wrote that
::_getFullyQualifiedClassName() method to use it in code generation.

the method has been corrected, usages have been corrected and the auto-
load related ClassLoaderWrapper is giving error and throws a
RuntimeException if a classname (as it must be FQCN in PHP autoloading)
is invalid.

This is especially important as there has been a bug back in very early re-
visions of PHP 5.3 where PHP itself prefixed with backslash (seems to be
a somewhat common mistake).

As much as the mistake might be common in certain circles, the more it is
important to clearly communicate at all points that passing a class-name
that is a FQCN must not start with a backslash to educate users to fully
grasp the concept of the FQCN.

Additional:

- Superfluous double-the-same-function-call removed.

- The Wrapper is a Decorator, Docblock speaks common language now.

- Only if the file was not found previously, the entry in the classmap
  will be NULLed. Otherwise an existing correct entry might be over-
  written.