bin/magento must return non zero on exceptions
When you use bin/magento in a deployment flow it must return a non zero
exitcode when there was an error or an exception. Now the exceptions
caught and as a result the exitcode is 0. In a script we assume exit
code 0 to be all clear everything went fine.

So we must fix this here.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>