Fix fwrite() truncation on resource
When using fwrite() operation on a resource, it is not a given to expect
that all data has been written in the operation.

This fix still checks for the error return value, but also if all bytes
have been written and if not all bytes have yet been written, the write
operation will continue until all bytes have been written *or* an error
occurs.

This is particularly important for operation on network streams.