Use PHP_VERSION_ID to check the minimum required version
This is a micro-optimisation, but it avoids the overhead of two function
calls to determine whether the minimum required version of PHP is present.

I have also removed the conditional around the a call to gc_collect_cycles
within the tests since it checks for version 5.3+, and the minimum version
required to run Magento 2 should already satisfy that condition.