fix multi storeview with lang path in base_uri
We have a website for example http://magento2.dev/

We have 3 stores,
locale nl_NL; uses same base_url as website
locale fr_FR; uses http://magento2.dev/fr/
locale en_US; uses http://magento2.dev/en/

when we go to the homepage and switch the language everything seems to
work.

But for example when we go to a category page outside the "default site"
url we get a 404.

This is normal because we find no match in the routing.

if our category page is /category/subcategory we match in the
UrlRewriteRouter to /catalog/categroy/view/id/x .

When we are in a storeview where there is /fr or /en in the beginning of
the path, we match nothing and result in a 404.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>