Add nl_BE and fr_BE languages for Belgium
For our Belgian customers it would be very nice if their languages were also supported. Although Belgians speak Dutch and French there is a profound difference between Dutch and Flemish (nl_BE) and between French and Walloon (fr_BE). They require different language packs.

Although nl_BE and fr_BE can be easily added with a plugin - which is my current solution - it's a small change to add it to the Magento code. I think the Belgians would be thankful.