Store, add launch plugin to remove storeview specifics
When a request is launched the area is deterined durning launch.
When we have storeviews with the same domain and an added path in the
url, all requests for that storeview will land in the area 'frontend'
while we could have requests for webapi_rest, ...

Before the cleaning of the store specific path was already done in the
request preprocessor but for other than frontend requests this pathInfo
update comes to late. So now the pathInfo is updated during launch.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>