Change value-assignment to a setValue call
Is there any reasons not to use Prototype's setValue call instead of assignment?
Plain assignment breaks in case of a multiselect element. 
In case of multiselect, and an array in
rowData.column_values[rowInputElementNames[i]]
the first element of that array will be taken and assigned as the only selected value on a multiselect, which is clearly a broken behaviour.
Prototype's setValue call will handle this case properly without breaking compatibility.