Cast string SKUs back to string
Array keys cast to integer on string SKUs that could represent a PHP integer, however the Mysql comparison is string based for which - to prevent a cast in Mysql - it needs string quotes. These string quotes then need again the $sku as string.

This is fixed by casting to string.

Background: If you import many records with SKU that are numeric strings, Mysql performance is impacted negatively.