Bug fix: enable adding more than one image in editors
Steps to reproduce:
- Edit a CMS page
- Add an image
- Add a second image

Result:
- After the image selection modal is closed a grey layover remains visible, but the tinymce image editor does not show anymore

Expected result:
- Tinymce editor is shown again after the image selection modal closes and I can add/edit my image

The error is that in the `closed: function() {` the variable `frameDialog` references to the first opened tinymce editor element. tinymce creates a new element every time you open it with a unique id, so the one you try to show the second time does not exist anymore.
I fixed it by declaring the frameDialog globally but feel free to solve this scope issue an other way if that's more conventional.