Changed constructor to use an interface
The front controller only accepts an HTTP response object. The front controller now has a direct dependency on the Response\Http class, which is not correct. The front controller should not care which response object will be injected only that it will adhire the ResponseInterface. So changed this constructor to accept a ResponseInterface which in di.xml is being replaced by the Response\Http class.