Move cache to instance of price box widget
When caches are shared on the class of the price box widget, there can
be a case where price boxes are displayed on the page which do not share
the same values that should be stored in that cache. For instance,
upsells and related products can have price boxes in them, and they will
overwrite the cache of the price box on the product page. This change
moves the cache off of the class itself and onto the instance so
instances of price box widgets don't interfere with one another.