Add de_LU and fr_LU languages for Luxembourg
For our Luxembourgian customers it would be very nice if their languages were also supported.

Although de_LU and fr_LU can be easily added with a plugin - which is my current solution - it's a small change to add it to the Magento code. I think the Luxembourgians would be thankful.