Initialize the attribute data for the entity type
Traditionally we wait for the first attribute of an entity in order to automatically initialize the array.  This is normally fine - however, when an entity supports EAV attributes but has none, this means that the attributeData is never initialized.

This causes issues on Line 512 when array_keys() is called on the array - rendering a PHP FATAL ERROR instead of an empty array.