Do not try to parse comments from cross-domain iframes
In the page-cache.js module the comments jQuery function tries to parse
comments from each element in the DOM. This fails when it tries to parse any
iframe elements that have a src attribute from a different domain.

This change skips any iframe element that does not belong to the current
domain.