Updated Swagger output to fix operation parameter name output when parameter "in" is "body"
This was previously being output as a string "$body" which appears to have been a workaround for a now-fixed swagger-codegen bug (#630). The convention appears to be name "body" when using in "body" elsewhere.

Other parameters in Magento's swagger file are not pre-fixed with a dollar sign. Removing this would prevent variable variable assignments when using php-based Swagger / Open API code generators like janephp/openapi.