Fix invalid asset resolution with numeric theme_id
There are cases in which the theme *ID* is passed to the asset resolution, but the processing in this case doesn't handle numbers correctly. 

This bug was hidden this much time because the validation of the theme model is done incorrectly, assuming it can sometimes be falsy, which is not the case, as the model is a result of `$themeCollection->getFirstItem()`, which will always be an object (so truthy).