Allow USPS Shipping Methods Without ExtraServices
According to the USPS API `<ExtraServices>` can show up as an empty element, like so: `<ExtraServices/>`. In fact, this is causing an issue with "Priority Mail International" shipments to the UK, because USPS doesn't offer any extra services for that shipping method. They do however add extra services to Priority Mail International packages sent to AU.